Refactor QC charge generation into new submodule
