+++ openff/recharge/charges/__init__.py
@@ -1,5 +1,1 @@
 """Generate partial charges, charge corrections and virtual sites for molecules"""
-
-from openff.recharge.charges._charges import ChargeGenerator, ChargeSettings
-
-__all__ = ["ChargeGenerator", "ChargeSettings"]
